Create Car_Game
An interactive car game where users type "Start" to start the car, "Stop" to stop the car, "Help" to bring up a help menu, and "Quit" to terminate the program. The program is not case sensitive and will present error messages if repeat commands are entered. For example, if the user types "Start" while the car has been already started, the program will let the user know that the car has already been started. 

This program incorporates loops, if statements, variables, booleans, and operators.